UPD: Typed card room
