+++ package.json
@@ -40,6 +40,7 @@
     "@types/http-proxy-middleware": "^0.19.3",
     "@types/lodash": "^4.14.161",
     "@types/prop-types": "^15.7.3",
+    "@types/react-router-dom": "^5.1.5",
     "@types/react-transition-group": "^4.4.0",
     "@types/styled-components": "^5.1.3",
     "@typescript-eslint/eslint-plugin": "^4.3.0",
